Rename _base.py file to core.py and update imports
